@property and @setter created for gyro_odr and mag_odr
The value that is AND with the reg for property and setter are assumed to be the 111 for the block this reg is replacing. Also disabled accel odr for now since if both gyro and accel are activated, then gyro frequencies are the same for the accel. Assuming that self. value doesn't need to be set for mag or gyro since it's not used for multiplication anywhere. That may mean that the _LSM9DS1_MAG and _LSM9DS1_GYRO constants aren't necessary.